Fix detection of chunk start
There seems to have been a change from ````r` to ```` r` that was throwing off the `chunk_start` regex.  This commit adds ` ?` to the regex, which allows for an optional space.  Relatedly, I have also removed some of the use of chunk references within litr's index.Rmd.  This can be put back later.